+++ super_dave/changelog.md
@@ -1,6 +1,9 @@
 ﻿# 0.0.7
 
 ## Features
+- Added infinite crab traps option.
+- Added infinite salvage drones option.
+- Added droneless large pickup option.
 
 ## Bug Fixes
 - Sleep Aura no longer hits already sleeping fish.